Add files via upload
Cybersecurity Keylogger Project:

Introduction:
Welcome to the Cybersecurity Keylogger Project! This repository hosts an open-source keylogger designed for educational and ethical purposes only. The primary goal of this project is to raise awareness about cybersecurity vulnerabilities and promote responsible use of information security tools.

Disclaimer: The use of keyloggers for any unauthorized or malicious activities is strictly prohibited. Ensure compliance with local laws and ethical guidelines before using this tool. The project contributors and maintainers are not responsible for any misuse of the software.

Features
Educational Purpose: This keylogger is intended for educational and research purposes to understand and enhance cybersecurity defenses.
Responsible Use: Users are expected to adhere to ethical guidelines and legal regulations when deploying this tool.
Customizable: The code is open-source, allowing users to customize and enhance the keylogger based on their specific needs.